+++ headers/HitableList.hpp
@@ -35,10 +35,10 @@ public:
 		return true;
 	}
 
-	bool hit(const Ray& r, float t_min, float t_max, HitRecord& rec) const override{
+	bool hit(const Ray& r, double t_min, double t_max, HitRecord& rec) const override{
 		HitRecord temp_rec;
 		bool hit_anything = false;
-		float closest_so_far = t_max;
+		double closest_so_far = t_max;
 		for (int i = 0; i < list_size; i++){
 			if (list[i]->hit(r, t_min, closest_so_far, temp_rec)){
 				hit_anything = true;
