+++ headers/Material.hpp
@@ -14,10 +14,10 @@ Vec3 reflect(const Vec3 &v, const Vec3 &n){
 	return v - 2*Vec3::dot(v, n)*n;
 }
 
-bool refract(const Vec3& v, const Vec3& n, float ni_over_nt, Vec3& refracted){
+bool refract(const Vec3& v, const Vec3& n, double ni_over_nt, Vec3& refracted){
 	Vec3 uv = Vec3::unit_vector(v);
-	float dt = Vec3::dot(uv, n);
-	float discriminant = 1.0 - ni_over_nt*ni_over_nt*(1-dt*dt);
+	double dt = Vec3::dot(uv, n);
+	double discriminant = 1.0 - ni_over_nt*ni_over_nt*(1-dt*dt);
 	if (discriminant > 0){
 		refracted = ni_over_nt*(uv - n*dt) - n*sqrt(discriminant);
 		return true;
@@ -26,15 +26,15 @@ bool refract(const Vec3& v, const Vec3& n, float ni_over_nt, Vec3& refracted){
 		return false;
 }
 
-float schlick(float cosine, float ref_idx){
-	float r0 = (1-ref_idx)/(1+ref_idx);
+double schlick(double cosine, double ref_idx){
+	double r0 = (1-ref_idx)/(1+ref_idx);
 	r0 = r0*r0;
 	return r0 + (1-r0)*pow((1-cosine),5);
 }
 
 class Texture{
 public:
-	virtual Vec3 value(float u, float v, const Vec3& p) const = 0;
+	virtual Vec3 value(double u, double v, const Vec3& p) const = 0;
 };
 
 class Material{
@@ -69,9 +69,9 @@ class Metal: public Material{
 public:
 	
 	Vec3 albedo;
-	float fuzz;
+	double fuzz;
 
-	Metal(const Vec3& a, float f){
+	Metal(const Vec3& a, double f){
 		albedo = a;
 		if (f < 0) fuzz = 0;
 		else if (f < 1) fuzz = f;
@@ -93,8 +93,8 @@ public:
 
 	Vec3 transparency;
 
-	float ref_idx;
-	Dielectric(float ri){
+	double ref_idx;
+	Dielectric(double ri){
 		transparency = Vec3(1-random()*0.2,1-random()*0.2,1-random()*0.2);
 		ref_idx = ri;
 	}
@@ -103,12 +103,12 @@ public:
 	virtual bool scatter(const Ray &r_in, const HitRecord &rec, Vec3 &attenuation, Ray &scattered) const{
 		Vec3 outward_normal;
 		Vec3 reflected = reflect(r_in.direction(), rec.normal);
-		float ni_over_nt;
+		double ni_over_nt;
 		attenuation = transparency;
 		Vec3 refracted;
-		float reflect_prob;
-		float cosine;
-		float dot_d_n = Vec3::dot(r_in.direction(), rec.normal);
+		double reflect_prob;
+		double cosine;
+		double dot_d_n = Vec3::dot(r_in.direction(), rec.normal);
 		if (dot_d_n > 0){
 			outward_normal = -rec.normal;
 			ni_over_nt = ref_idx;
